Refactor Connection, Job, Result to modules under api namespace
