+++ strawberryfields/engine.py
@@ -33,110 +33,13 @@ import requests
 from strawberryfields.configuration import DEFAULT_CONFIG
 from strawberryfields.io import to_blackbird
 from strawberryfields.program import Program
+from strawberryfields.api import Connection, Job, JobStatus, Result
 
 from .backends import load_backend
 from .backends.base import BaseBackend, NotApplicableError
 
 # for automodapi, do not include the classes that should appear under the top-level strawberryfields namespace
 __all__ = ["Result", "BaseEngine", "LocalEngine", "Connection"]
-
-
-class Result:
-    """Result of a quantum computation.
-
-    Represents the results of the execution of a quantum program on a local or
-    remote backend.
-
-    The returned :class:`~Result` object provides several useful properties
-    for accessing the results of your program execution:
-
-    * ``results.state``: The quantum state object contains details and methods
-      for manipulation of the final circuit state. Not available for remote
-      backends. See :doc:`/introduction/states` for more information regarding available
-      state methods.
-
-    * ``results.samples``: Measurement samples from any measurements performed.
-
-    **Example:**
-
-    The following examples run an existing Strawberry Fields
-    quantum :class:`~.Program` on the Gaussian engine to get
-    a results object.
-
-    Using this results object, the measurement samples
-    can be returned, as well as quantum state information.
-
-    >>> eng = sf.Engine("gaussian")
-    >>> results = eng.run(prog)
-    >>> print(results)
-    Result: 3 subsystems
-        state: <GaussianState: num_modes=3, pure=True, hbar=2>
-        samples: [0, 0, 0]
-    >>> results.samples
-    [0, 0, 0]
-    >>> results.state.is_pure()
-    True
-
-    .. note::
-
-        Only local simulators will return a state object. Remote
-        simulators and hardware backends will return
-        measurement samples  (:attr:`Result.samples`),
-        but the return value of ``Result.state`` will be ``None``.
-    """
-
-    def __init__(self, samples, is_stateful=True):
-        self._state = None
-        self._is_stateful = is_stateful
-
-        # ``samples`` arrives as a list of arrays, need to convert here to a multidimensional array
-        if len(np.shape(samples)) > 1:
-            samples = np.stack(samples, 1)
-        self._samples = samples
-
-    @property
-    def samples(self):
-        """Measurement samples.
-
-        Returned measurement samples will have shape ``(modes,)``. If multiple
-        shots are requested during execution, the returned measurement samples
-        will instead have shape ``(shots, modes)``.
-
-        Returns:
-            array[array[float, int]]: measurement samples returned from
-            program execution
-        """
-        return self._samples
-
-    @property
-    def state(self):
-        """The quantum state object.
-
-        The quantum state object contains details and methods
-        for manipulation of the final circuit state.
-
-        See :doc:`/introduction/states` for more information regarding available
-        state methods.
-
-        .. note::
-
-            Only local simulators will return a state object. Remote
-            simulators and hardware backends will return
-            measurement samples (:attr:`Result.samples`),
-            but the return value of ``Result.state`` will be ``None``.
-
-        Returns:
-            BaseState: quantum state returned from program execution
-        """
-        if not self._is_stateful:
-            raise AttributeError("The state is undefined for a stateless computation.")
-        return self._state
-
-    def __str__(self):
-        """String representation."""
-        return "Result: {} subsystems, state: {}\n samples: {}".format(
-            len(self.samples), self.state, self.samples
-        )
 
 
 class BaseEngine(abc.ABC):
@@ -544,374 +447,8 @@ class LocalEngine(BaseEngine):
         return result
 
 
-class RequestFailedError(Exception):
-    """Raised when a request to the remote platform returns an error response."""
-
-
-class InvalidJobOperationError(Exception):
-    """Raised when an invalid operation is performed on a job."""
-
-
 class JobFailedError(Exception):
     """Raised when a remote job enters a 'failed' status."""
-
-
-class JobStatus(enum.Enum):
-    """Represents the status of a remote job.
-
-    This class maps a set of job statuses to the string representations returned by the
-    remote platform.
-    """
-
-    OPEN = "open"
-    QUEUED = "queued"
-    CANCELLED = "cancelled"
-    COMPLETED = "complete"
-    FAILED = "failed"
-
-    @property
-    def is_final(self) -> bool:
-        """Checks if this status represents a final and immutable state.
-
-        This method is primarily used to determine if an operation is valid for a given
-        status.
-
-        Returns:
-            bool: ``True`` if the job status is final, and ``False`` otherwise
-        """
-        return self in (JobStatus.CANCELLED, JobStatus.COMPLETED, JobStatus.FAILED)
-
-
-class Job:
-    """Represents a remote job that can be queried for its status or result.
-
-    This object should typically not be instantiated directly, but returned by an
-    ``Engine`` or ``Connection`` when a job is run.
-
-    Args:
-        id_ (str): the job ID
-        status (strawberryfields.engine.JobStatus): the job status
-        connection (strawberryfields.engine.Connection): the connection over which the
-            job is managed
-    """
-
-    def __init__(self, id_: str, status: JobStatus, connection: "Connection"):
-        self._id = id_
-        self._status = status
-        self._connection = connection
-        self._result = None
-
-    @property
-    def id(self) -> str:
-        """The job ID.
-
-        Returns:
-            str: the job ID
-        """
-        return self._id
-
-    @property
-    def status(self) -> JobStatus:
-        """The job status.
-
-        Returns:
-            strawberryfields.engine.JobStatus: the job status
-        """
-        return self._status
-
-    @property
-    def result(self) -> Result:
-        """The job result.
-
-        This is only defined for completed jobs, and raises an exception for any other
-        status.
-
-        Returns:
-            strawberryfields.engine.Result: the job result
-        """
-        if self.status != JobStatus.COMPLETED:
-            raise AttributeError(
-                "The result is undefined for jobs that are not completed "
-                "(current status: {})".format(self.status.value)
-            )
-        return self._result
-
-    def refresh(self):
-        """Refreshes the status of the job, along with the job result if the job is
-        newly completed.
-
-        Only an open or queued job can be refreshed; an exception is raised otherwise.
-        """
-        if self.status.is_final:
-            raise InvalidJobOperationError(
-                "A {} job cannot be refreshed".format(self.status.value)
-            )
-        self._status = self._connection.get_job_status(self.id)
-        if self._status == JobStatus.COMPLETED:
-            self._result = self._connection.get_job_result(self.id)
-
-    def cancel(self):
-        """Cancels the job.
-
-        Only an open or queued job can be cancelled; an exception is raised otherwise.
-        """
-        if self.status.is_final:
-            raise InvalidJobOperationError(
-                "A {} job cannot be cancelled".format(self.status.value)
-            )
-        self._connection.cancel_job(self.id)
-
-    def __repr__(self):
-        return "<{}: id={}, status={}>".format(
-            self.__class__.__name__, self.id, self.status.value
-        )
-
-    def __str__(self):
-        return self.__repr__()
-
-
-class Connection:
-    """Manages remote connections to the remote job execution platform and exposes
-    various job operations.
-
-    For basic usage, it is not necessary to manually instantiate this object; the user
-    is encouraged to use the higher-level interface provided by :class:`~StarshipEngine`.
-
-    **Example:**
-
-    The following example instantiates a :class:`~Connection` for a given API
-    authentication token, tests the connection, submits a new job, and makes requests
-    for a single or multiple existing jobs.
-
-    .. code-block:: python
-
-        connection = Connection(token="abc")
-
-        # Ping the remote server
-        success = connection.ping()
-        # True if successful, or False if cannot connect or not authenticated
-
-        # Submit a new job
-        job = connection.create_job("chip2", program, shots=123)
-        job  # <Job: id=d177cbf5-1816-4779-802f-cef2c117dc1a, ...>
-
-        # Get all jobs submitted for this token
-        jobs = connection.get_all_jobs()
-        jobs  # [<Job: id=d177cbf5-1816-4779-802f-cef2c117dc1a, ...>, ...]
-
-        # Get a specific job by ID
-        job = connection.get_job("59a1c0b1-c6a7-4f9b-ae37-0ac5eec9c413")
-        job  # <Job: id=59a1c0b1-c6a7-4f9b-ae37-0ac5eec9c413, ...>
-
-    Args:
-        token (str): the API authentication token
-        host (str): the hostname of the remote platform
-        port (int): the port to connect to on the remote host
-        use_ssl (bool): whether to use SSL for the connection
-    """
-
-    MAX_JOBS_REQUESTED = 100
-    JOB_TIMESTAMP_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
-
-    # pylint: disable=bad-continuation
-    # See: https://github.com/PyCQA/pylint/issues/289
-    def __init__(
-        self,
-        token: str,
-        host: str = "platform.strawberryfields.ai",
-        port: int = 443,
-        use_ssl: bool = True,
-    ):
-        self._token = token
-        self._host = host
-        self._port = port
-        self._use_ssl = use_ssl
-
-        self._base_url = "http{}://{}:{}".format(
-            "s" if self.use_ssl else "", self.host, self.port
-        )
-        self._headers = {"Authorization": self.token}
-
-    @property
-    def token(self) -> str:
-        """The API authentication token.
-
-        Returns:
-            str: the authentication token
-        """
-        return self._token
-
-    @property
-    def host(self) -> str:
-        """The host for the remote platform.
-
-        Returns:
-            str: the hostname
-        """
-        return self._host
-
-    @property
-    def port(self) -> int:
-        """The port to connect to on the remote host.
-
-        Returns:
-            int: the port number
-        """
-        return self._port
-
-    @property
-    def use_ssl(self) -> bool:
-        """Whether to use SSL for the connection.
-
-        Returns:
-            bool: ``True`` if SSL should be used, and ``False`` otherwise
-        """
-        return self._use_ssl
-
-    def create_job(self, target: str, program: Program, shots: int) -> Job:
-        """Creates a job with the given circuit.
-
-        Args:
-            target (str): the target device
-            program (strawberryfields.Program): the quantum circuit
-            shots (int): the number of shots
-
-        Returns:
-            strawberryfields.engine.Job: the created job
-        """
-        # Serialize a blackbird circuit for network transmission
-        bb = to_blackbird(program)
-        # pylint: disable=protected-access
-        bb._target["name"] = target
-        # pylint: disable=protected-access
-        bb._target["options"] = {"shots": shots}
-        circuit = bb.serialize()
-
-        path = "/jobs"
-        response = requests.post(
-            self._url(path),
-            headers=self._headers,
-            data=json.dumps({"circuit": circuit}),
-        )
-        if response.status_code == 201:
-            return Job(
-                id_=response.json()["id"],
-                status=JobStatus(response.json()["status"]),
-                connection=self,
-            )
-        raise RequestFailedError(
-            "Job creation failed: {}".format(self._format_error_message(response))
-        )
-
-    def get_all_jobs(self, after: datetime = datetime(1970, 1, 1)) -> List[Job]:
-        """Gets a list of jobs created by the user, optionally filtered by datetime.
-
-        A maximum of the 100 most recent jobs are returned.
-
-        Args:
-            after (datetime.datetime): if provided, only jobs more recently created
-                then ``after`` are returned
-
-        Returns:
-            List[strawberryfields.engine.Job]: the jobs
-        """
-        raise NotImplementedError("This feature is not yet implemented")
-
-    def get_job(self, job_id: str) -> Job:
-        """Gets a job.
-
-        Args:
-            job_id (str): the job ID
-
-        Returns:
-            strawberryfields.engine.Job: the job
-        """
-        path = "/jobs/{}".format(job_id)
-        response = requests.get(self._url(path), headers=self._headers)
-        if response.status_code == 200:
-            return Job(
-                id_=response.json()["id"],
-                status=JobStatus(response.json()["status"]),
-                connection=self,
-            )
-        raise RequestFailedError(self._format_error_message(response))
-
-    def get_job_status(self, job_id: str) -> JobStatus:
-        """Returns the status of a job.
-
-        Args:
-            job_id (str): the job ID
-
-        Returns:
-            strawberryfields.engine.JobStatus: the job status
-        """
-        return JobStatus(self.get_job(job_id).status)
-
-    def get_job_result(self, job_id: str) -> Result:
-        """Returns the result of a job.
-
-        Args:
-            job_id (str): the job ID
-
-        Returns:
-            strawberryfields.engine.Result: the job result
-        """
-        path = "/jobs/{}/result".format(job_id)
-        response = requests.get(
-            self._url(path), headers={"Accept": "application/x-numpy", **self._headers},
-        )
-        if response.status_code == 200:
-            # Read the numpy binary data in the payload into memory
-            with io.BytesIO() as buf:
-                buf.write(response.content)
-                buf.seek(0)
-                samples = np.load(buf)
-            return Result(samples, is_stateful=False)
-        raise RequestFailedError(self._format_error_message(response))
-
-    def cancel_job(self, job_id: str):
-        """Cancels a job.
-
-        Args:
-            job_id (str): the job ID
-        """
-        path = "/jobs/{}".format(job_id)
-        response = requests.patch(
-            self._url(path),
-            headers=self._headers,
-            data={"status", JobStatus.CANCELLED.value},
-        )
-        if response.status_code == 204:
-            return
-        raise RequestFailedError(self._format_error_message(response))
-
-    def ping(self) -> bool:
-        """Tests the connection to the remote backend.
-
-        Returns:
-            bool: ``True`` if the connection is successful, and ``False`` otherwise
-        """
-        path = "/healthz"
-        response = requests.get(self._url(path), headers=self._headers)
-        return response.status_code == 200
-
-    def _url(self, path: str) -> str:
-        return self._base_url + path
-
-    @staticmethod
-    def _format_error_message(response: requests.Response) -> str:
-        body = response.json()
-        return "{} ({}): {}".format(
-            body.get("status_code", ""), body.get("code", ""), body.get("detail", "")
-        )
-
-    def __repr__(self):
-        return "<{}: token={}, host={}>".format(
-            self.__class__.__name__, self.token, self.host
-        )
-
-    def __str__(self):
-        return self.__repr__()
 
 
 class StarshipEngine:
